telegram_youtube_bot/requirements.txt
python-telegram-bot==20.7
musicbrainzngs
yt-dlp
